requirements.txt
aiohttp
cinemagoer
cloudscraper
colour
cowpy
emoji==1.7.0
fonttools
geopy
gitpython
glitch_this
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gtts
hachoir
heroku3
html-telegraph-poster
httpx[http2]
humanize
jikanpy
justwatch
lottie
lyricsgenius
markdown
motor
Pillow
prettytable
psutil
psycopg2
pyfiglet
PyGithub
pygments
pylast
pymediainfo
PyMuPDF
pySmartDL
python-barcode
pytz
qrcode
requests
selenium
setuptools
ShazamAPI
spamwatch
speedtest-cli
sqlalchemy-json
sqlalchemy==1.3.23
telegraph==2.1.0
tgcrypto
ujson
urlextract
validators
vcsi
wand
wget
yt_dlp

git+https://github.com/Jisan09/search-engine-parser
git+https://github.com/Jisan09/Telethon@beta
git+https://github.com/goldsmith/Wikipedia
git+https://github.com/sandy1709/py-googletrans
git+https://github.com/alexmercerind/youtube-search-python
